Atom Entry class

This class models an atom Entry

Format

R6Class object.

Value

Object of R6Class for modelling an Atom Entry

Details

AtomEntry

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

atom4R::atom4RLogger -> atom4R::AtomAbstractObject -> AtomEntry

Public fields

id

identifier

updated

Update date/time

published

Publication date/time

title

Title

summary

Summary

rights

Rights

source

Source

author

Author(s)

contributor

Contributor(s)

category

Category

content

Content

Methods

Public methods

AtomEntry$new()

AtomEntry$setId()

AtomEntry$setUpdated()

AtomEntry$setPublished()

AtomEntry$setTitle()

AtomEntry$setSummary()

AtomEntry$setRights()

AtomEntry$setSource()

AtomEntry$addAuthor()

AtomEntry$delAuthor()

AtomEntry$addContributor()

AtomEntry$delContributor()

AtomEntry$addCategory()

AtomEntry$delCategory()

AtomEntry$addLink()

AtomEntry$delLink()

AtomEntry$setContent()

AtomEntry$clone()

Inherited methods

atom4R::atom4RLogger$ERROR()

atom4R::atom4RLogger$INFO()

atom4R::atom4RLogger$WARN()

atom4R::atom4RLogger$logger()

atom4R::AtomAbstractObject$addListElement()

atom4R::AtomAbstractObject$contains()

atom4R::AtomAbstractObject$createElement()

atom4R::AtomAbstractObject$decode()

atom4R::AtomAbstractObject$delListElement()

atom4R::AtomAbstractObject$encode()

atom4R::AtomAbstractObject$getClass()

atom4R::AtomAbstractObject$getClassName()

atom4R::AtomAbstractObject$getNamespace()

atom4R::AtomAbstractObject$getNamespaceDefinition()

atom4R::AtomAbstractObject$getRootElement()

atom4R::AtomAbstractObject$getXmlElement()

atom4R::AtomAbstractObject$isDocument()

atom4R::AtomAbstractObject$isFieldInheritedFrom()

atom4R::AtomAbstractObject$print()

atom4R::AtomAbstractObject$save()

atom4R::AtomAbstractObject$setIsDocument()

atom4R::AtomAbstractObject$validate()

Method new()

Initializes an AtomEntry

Usage

AtomEntry$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class from XML

Method setId()

Set ID

Usage

AtomEntry$setId(id)

Arguments

id

id

Method setUpdated()

Set updated date

Usage

AtomEntry$setUpdated(updated)

Arguments

updated

object of class Date or POSIXt

Method setPublished()

Set published date

Usage

AtomEntry$setPublished(published)

Arguments

published

object of class Date or POSIXt

Method setTitle()

Set title

Usage

AtomEntry$setTitle(title, type = "text")

Arguments

title

title

type

type. Default is "text"

Method setSummary()

Set summary

Usage

AtomEntry$setSummary(summary, type = "text")

Arguments

summary

summary

type

type. Default is "text"

Method setRights()

Set rights

Usage

AtomEntry$setRights(rights, type = "text")

Arguments

rights

rights

type

type. Default is "text"

Method setSource()

Set source

Usage

AtomEntry$setSource(source, type = "text")

Arguments

source

source

type

type. Default is "text"

Method addAuthor()

Adds author

Usage

AtomEntry$addAuthor(author)

Arguments

author

object of class AtomAuthor

Returns

TRUE if added, FALSE otherwise

Method delAuthor()

Deletes author

Usage

AtomEntry$delAuthor(author)

Arguments

author

object of class AtomAuthor

Returns

TRUE if deleted, FALSE otherwise

Method addContributor()

Adds contributor

Usage

AtomEntry$addContributor(contributor)

Arguments

contributor

object of class AtomContributor

Returns

TRUE if added, FALSE otherwise

Method delContributor()

Deletes contributor

Usage

AtomEntry$delContributor(contributor)

Arguments

contributor

object of class AtomContributor

Returns

TRUE if deleted, FALSE otherwise

Method addCategory()

Adds category

Usage

AtomEntry$addCategory(value, term, scheme = NULL, label = NULL)

Arguments

value

value

term

term

scheme

scheme

label

label

Returns

TRUE if added, FALSE otherwise

Method delCategory()

Deletes category

Usage

AtomEntry$delCategory(value, term, scheme = NULL, label = NULL)

Arguments

value

value

term

term

scheme

scheme

label

label

Returns

TRUE if deleted, FALSE otherwise

Method addLink()

Adds link

Usage

AtomEntry$addLink(link, rel = "alternate", type = "text/html")

Arguments

link

link

rel

relation. Default is "alternate"

type

type. Default is "text/html"

Returns

TRUE if added, FALSE otherwise

Method delLink()

Deletes link

Usage

AtomEntry$delLink(link, rel = "alternate", type = "text/html")

Arguments

link

link

rel

relation. Default is "alternate"

type

type. Default is "text/html"

Returns

TRUE if deleted, FALSE otherwise

Method setContent()

Set content

Usage

AtomEntry$setContent(content)

Arguments

content

content

Method clone()

The objects of this class are cloneable with this method.

Usage

AtomEntry$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 # \dontrun{
  #encoding
  atom <- AtomEntry$new()
  atom$setId("my-atom-entry")
  atom$setTitle("My Atom feed entry")
  atom$setSummary("My Atom feed entry very comprehensive abstract")
  author1 <- AtomAuthor$new(
    name = "John Doe",
    uri = "http://www.atomxml.com/johndoe",
    email = "johndoe@atom4R.com"
  )
  atom$addAuthor(author1)
#> [1] TRUE
  author2 <- AtomAuthor$new(
    name = "John Doe's sister",
    uri = "http://www.atomxml.com/johndoesister",
    email = "johndoesister@atom4R.com"
  )
  atom$addAuthor(author2)
#> [1] TRUE
  contrib1 <- AtomContributor$new(
    name = "Contrib1",
    uri = "http://www.atomxml.com/contrib1",
    email = "contrib1@atom4R.com"
  )
  atom$addContributor(contrib1)
#> [1] TRUE
  contrib2 <- AtomContributor$new(
    name = "Contrib2",
    uri = "http://www.atomxml.com/contrib2",
    email = "contrib2@atom4R.com"
  )
  atom$addContributor(contrib2)
#> [1] TRUE
  atom$addCategory("draft", "dataset")
#> [1] TRUE
  atom$addCategory("world", "spatial")
#> [1] TRUE
  atom$addCategory("fisheries", "domain")
#> [1] TRUE

  xml <- atom$encode()
#> [atom4R][WARN] Element '{http://www.w3.org/2005/Atom}category': Character content is not allowed, because the content type is empty at line 26. 
#> [atom4R][WARN] Element '{http://www.w3.org/2005/Atom}category': Character content is not allowed, because the content type is empty at line 27. 
#> [atom4R][WARN] Element '{http://www.w3.org/2005/Atom}category': Character content is not allowed, because the content type is empty at line 28. 
#> [atom4R][WARN] Object 'AtomEntry' is INVALID according to Atom XML schemas! 
 # }