Dublin Core Entry class

This class models an Dublin Core Entry

Format

R6Class object.

Value

Object of R6Class for modelling an Dublin Core Entry

Details

DCEntry

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

atom4R::atom4RLogger -> atom4R::AtomAbstractObject -> atom4R::AtomEntry -> DCEntry

Methods

Public methods

DCEntry$new()

DCEntry$addDCElement()

DCEntry$delDCElement()

DCEntry$setDCElements()

DCEntry$getDCElements()

DCEntry$getDCElementByValue()

DCEntry$addDCAbstract()

DCEntry$delDCAbstract()

DCEntry$setDCAbstracts()

DCEntry$getDCAbstracts()

DCEntry$addDCAccessRights()

DCEntry$delDCAccessRights()

DCEntry$setDCAccessRights()

DCEntry$getDCAccessRights()

DCEntry$addDCAccrualMethod()

DCEntry$delDCAccrualMethod()

DCEntry$setDCAccrualMethods()

DCEntry$getDCAccrualMethods()

DCEntry$addDCAccrualPeriodicity()

DCEntry$delDCAccrualPeriodicity()

DCEntry$setDCAccrualPeriodicities()

DCEntry$getDCAccrualPeriodicities()

DCEntry$addDCAccrualPolicy()

DCEntry$delDCAccrualPolicy()

DCEntry$setDCAccrualPolicies()

DCEntry$getDCAccrualPolicies()

DCEntry$addDCAlternative()

DCEntry$delDCAlternative()

DCEntry$setDCAlternatives()

DCEntry$getDCAlternatives()

DCEntry$addDCAudience()

DCEntry$delDCAudience()

DCEntry$setDCAudiences()

DCEntry$getDCAudiences()

DCEntry$addDCAvailable()

DCEntry$delDCAvailable()

DCEntry$setDCAvailables()

DCEntry$getDCAvailables()

DCEntry$addDCBibliographicCitation()

DCEntry$delDCBibliographicCitation()

DCEntry$setDCBibliographicCitations()

DCEntry$getDCBibliographicCitations()

DCEntry$addDCConformsTo()

DCEntry$delDCConformsTo()

DCEntry$setDCConformsTo()

DCEntry$getDCConformsTo()

DCEntry$addDCContributor()

DCEntry$delDCContributor()

DCEntry$setDCContributors()

DCEntry$getDCContributors()

DCEntry$addDCCoverage()

DCEntry$delDCCoverage()

DCEntry$setDCCoverages()

DCEntry$getDCCoverages()

DCEntry$addDCCreated()

DCEntry$delDCCreated()

DCEntry$addDCCreator()

DCEntry$delDCCreator()

DCEntry$setDCCreators()

DCEntry$getDCCreators()

DCEntry$addDCDate()

DCEntry$delDCDate()

DCEntry$setDCDates()

DCEntry$getDCDates()

DCEntry$addDCDateAccepted()

DCEntry$delDCDateAccepted()

DCEntry$addDCDateCopyrighted()

DCEntry$delDCDateCopyrighted()

DCEntry$addDCDateSubmitted()

DCEntry$delDCDateSubmitted()

DCEntry$addDCDescription()

DCEntry$delDCDescription()

DCEntry$setDCDescriptions()

DCEntry$getDCDescriptions()

DCEntry$addDCEducationalLevel()

DCEntry$delDCEducationalLevel()

DCEntry$setDCEducationalLevels()

DCEntry$getDCEducationalLevels()

DCEntry$addDCExtent()

DCEntry$delDCExtent()

DCEntry$setDCExtents()

DCEntry$getDCExtents()

DCEntry$addDCFormat()

DCEntry$delDCFormat()

DCEntry$setDCFormats()

DCEntry$getDCFormats()

DCEntry$addDCHasPart()

DCEntry$delDCHasPart()

DCEntry$setDCHasParts()

DCEntry$getDCHasParts()

DCEntry$addDCHasVersion()

DCEntry$delDCHasVersion()

DCEntry$setDCHasVersions()

DCEntry$getDCHasVersions()

DCEntry$addDCIdentifier()

DCEntry$delDCIdentifier()

DCEntry$setDCIdentifiers()

DCEntry$getDCIdentifiers()

DCEntry$addDCInstructionalMethod()

DCEntry$delDCInstructionalMethod()

DCEntry$setDCInstructionalMethods()

DCEntry$getDCInstructionalMethods()

DCEntry$addDCIsPartOf()

DCEntry$delDCIsPartOf()

DCEntry$setDCIsPartOf()

DCEntry$getDCIsPartOfs()

DCEntry$addDCIsReferencedBy()

DCEntry$delDCIsReferencedBy()

DCEntry$setDCIsReferencedBys()

DCEntry$getDCIsReferencedBys()

DCEntry$addDCIsReplacedBy()

DCEntry$delDCIsReplacedBy()

DCEntry$setDCIsReplacedBys()

DCEntry$getDCIsReplacedBys()

DCEntry$addDCIsRequiredBy()

DCEntry$delDCIsRequiredBy()

DCEntry$setDCIsRequiredBys()

DCEntry$getDCIsRequiredBys()

DCEntry$addDCIsVersionOf()

DCEntry$delDCIsVersionOf()

DCEntry$setDCIsVersionOfs()

DCEntry$getDCIsVersionOfs()

DCEntry$addDCIssued()

DCEntry$delDCIssued()

DCEntry$addDCLanguage()

DCEntry$delDCLanguage()

DCEntry$setDCLanguages()

DCEntry$getDCLanguages()

DCEntry$addDCLicense()

DCEntry$delDCLicense()

DCEntry$setDCLicenses()

DCEntry$getDCLicenses()

DCEntry$addDCMediator()

DCEntry$delDCMediator()

DCEntry$setDCMediators()

DCEntry$getDCMediators()

DCEntry$addDCMedium()

DCEntry$delDCMedium()

DCEntry$setDCMediums()

DCEntry$getDCMediums()

DCEntry$addDCModified()

DCEntry$delDCModified()

DCEntry$addDCProvenance()

DCEntry$delDCProvenance()

DCEntry$setDCProvenances()

DCEntry$getDCProvenances()

DCEntry$addDCPublisher()

DCEntry$delDCPublisher()

DCEntry$setDCPublishers()

DCEntry$getDCPublishers()

DCEntry$addDCReferences()

DCEntry$delDCReferences()

DCEntry$setDCReferences()

DCEntry$getDCReferences()

DCEntry$addDCRelation()

DCEntry$delDCRelation()

DCEntry$setDCRelations()

DCEntry$getDCRelations()

DCEntry$addDCReplaces()

DCEntry$delDCReplaces()

DCEntry$setDCReplaces()

DCEntry$getDCReplaces()

DCEntry$addDCRequires()

DCEntry$delDCRequires()

DCEntry$setDCRequires()

DCEntry$getDCRequires()

DCEntry$addDCRights()

DCEntry$delDCRights()

DCEntry$setDCRights()

DCEntry$getDCRights()

DCEntry$addDCRightsHolder()

DCEntry$delDCRightsHolder()

DCEntry$setDCRightsHolders()

DCEntry$getDCRightsHolders()

DCEntry$addDCSource()

DCEntry$delDCSource()

DCEntry$setDCSources()

DCEntry$getDCSources()

DCEntry$addDCSubject()

DCEntry$delDCSubject()

DCEntry$setDCSubjects()

DCEntry$getDCSubjects()

DCEntry$addDCTableOfContents()

DCEntry$delDCTableOfContents()

DCEntry$setDCTablesOfContents()

DCEntry$getDCTablesOfContent()

DCEntry$addDCTemporal()

DCEntry$delDCTemporal()

DCEntry$setDCTemporals()

DCEntry$getDCTemporals()

DCEntry$addDCTitle()

DCEntry$delDCTitle()

DCEntry$setDCTitles()

DCEntry$getDCTitles()

DCEntry$addDCType()

DCEntry$delDCType()

DCEntry$setDCTypes()

DCEntry$getDCTypes()

DCEntry$asDataFrame()

DCEntry$clone()

Inherited methods

atom4R::atom4RLogger$ERROR()

atom4R::atom4RLogger$INFO()

atom4R::atom4RLogger$WARN()

atom4R::atom4RLogger$logger()

atom4R::AtomAbstractObject$addListElement()

atom4R::AtomAbstractObject$contains()

atom4R::AtomAbstractObject$createElement()

atom4R::AtomAbstractObject$decode()

atom4R::AtomAbstractObject$delListElement()

atom4R::AtomAbstractObject$encode()

atom4R::AtomAbstractObject$getClass()

atom4R::AtomAbstractObject$getClassName()

atom4R::AtomAbstractObject$getNamespace()

atom4R::AtomAbstractObject$getNamespaceDefinition()

atom4R::AtomAbstractObject$getRootElement()

atom4R::AtomAbstractObject$getXmlElement()

atom4R::AtomAbstractObject$isDocument()

atom4R::AtomAbstractObject$isFieldInheritedFrom()

atom4R::AtomAbstractObject$print()

atom4R::AtomAbstractObject$save()

atom4R::AtomAbstractObject$setIsDocument()

atom4R::AtomAbstractObject$validate()

atom4R::AtomEntry$addAuthor()

atom4R::AtomEntry$addCategory()

atom4R::AtomEntry$addContributor()

atom4R::AtomEntry$addLink()

atom4R::AtomEntry$delAuthor()

atom4R::AtomEntry$delCategory()

atom4R::AtomEntry$delContributor()

atom4R::AtomEntry$delLink()

atom4R::AtomEntry$setContent()

atom4R::AtomEntry$setId()

atom4R::AtomEntry$setPublished()

atom4R::AtomEntry$setRights()

atom4R::AtomEntry$setSource()

atom4R::AtomEntry$setSummary()

atom4R::AtomEntry$setTitle()

atom4R::AtomEntry$setUpdated()

Method new()

Initializes an object of class DCEntry

Usage

DCEntry$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class from XML

Method addDCElement()

Adds a Dublin Core element

Usage

DCEntry$addDCElement(term, value, extended = FALSE)

Arguments

term

term

value

value

extended

extended. Default is FALSE

Returns

TRUE if added, FALSE otherwise

Method delDCElement()

Deletes a Dublin Core element

Usage

DCEntry$delDCElement(term, value)

Arguments

term

term

value

value

Returns

TRUE if deleted, FALSE otherwise

Method setDCElements()

Set a list of DC elements

Usage

DCEntry$setDCElements(term, values)

Arguments

term

term

values

vector of values

Method getDCElements()

Get a list of DC elements

Usage

DCEntry$getDCElements(term)

Arguments

term

term

Returns

a list of objects extending DCElement

Method getDCElementByValue()

Get a DC element by value

Usage

DCEntry$getDCElementByValue(term, value)

Arguments

term

term

value

value

Method addDCAbstract()

Adds DC abstract

Usage

DCEntry$addDCAbstract(abstract)

Arguments

abstract

object of class DCAbstract or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAbstract()

Deletes DC abstract

Usage

DCEntry$delDCAbstract(abstract)

Arguments

abstract

object of class DCAbstract or vector of class character and length 1

Method setDCAbstracts()

Set DC abstracts

Usage

DCEntry$setDCAbstracts(abstracts)

Arguments

abstracts

abstracts, vector of class character

Method getDCAbstracts()

Get DC abstracts

Usage

DCEntry$getDCAbstracts()

Returns

a list of objects of class DCAbstract

Method addDCAccessRights()

Adds DC access rights

Usage

DCEntry$addDCAccessRights(accessRights)

Arguments

accessRights

object of class DCAccessRights or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccessRights()

Deletes DC access rights

Usage

DCEntry$delDCAccessRights(accessRights)

Arguments

accessRights

object of class DCAccessRights or vector of class character and length 1

Method setDCAccessRights()

Set access rights

Usage

DCEntry$setDCAccessRights(accessRights)

Arguments

accessRights

vector of class character

Method getDCAccessRights()

Get DC access rights

Usage

DCEntry$getDCAccessRights()

Returns

a list of objects of class DCAccessRights

Method addDCAccrualMethod()

Adds DC accrual method

Usage

DCEntry$addDCAccrualMethod(accrualMethod)

Arguments

accrualMethod

object of class DCAccrualMethod or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualMethod()

Deletes DC accrual method

Usage

DCEntry$delDCAccrualMethod(accrualMethod)

Arguments

accrualMethod

object of class DCAccrualMethod or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAccrualMethods()

Set DC accrual method

Usage

DCEntry$setDCAccrualMethods(accrualMethods)

Arguments

accrualMethods

vector of class character

Method getDCAccrualMethods()

Get DC accrual method

Usage

DCEntry$getDCAccrualMethods()

Returns

a list of objects of class DCAccrualMethod

Method addDCAccrualPeriodicity()

Adds DC accrual periodicity

Usage

DCEntry$addDCAccrualPeriodicity(accrualPeriodicity)

Arguments

accrualPeriodicity

object of class DCAccrualPeriodicity or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualPeriodicity()

Deletes DC accrual periodicity

Usage

DCEntry$delDCAccrualPeriodicity(accrualPeriodicity)

Arguments

accrualPeriodicity

object of class DCAccrualPeriodicity or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAccrualPeriodicities()

Set DC accrual periodicities

Usage

DCEntry$setDCAccrualPeriodicities(accrualPeriodicities)

Arguments

accrualPeriodicities

vector of class character

Method getDCAccrualPeriodicities()

Get DC accrual periodicities

Usage

DCEntry$getDCAccrualPeriodicities()

Returns

a list of objects of class DCAccrualPeriodicity

Method addDCAccrualPolicy()

Adds DC accrual policy

Usage

DCEntry$addDCAccrualPolicy(accrualPolicy)

Arguments

accrualPolicy

object of class DCAccrualPolicy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualPolicy()

Deletes DC accrual policy

Usage

DCEntry$delDCAccrualPolicy(accrualPolicy)

Arguments

accrualPolicy

object of class DCAccrualPolicy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAccrualPolicies()

Set DC accrual policies

Usage

DCEntry$setDCAccrualPolicies(accrualPolicies)

Arguments

accrualPolicies

vector of class character

Method getDCAccrualPolicies()

Get DC accrual policies

Usage

DCEntry$getDCAccrualPolicies()

Returns

a list of objects of class DCAccrualPolicy

Method addDCAlternative()

Adds DC alternative

Usage

DCEntry$addDCAlternative(alternative)

Arguments

alternative

object of class DCAlternative or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAlternative()

Deletes DC alternative

Usage

DCEntry$delDCAlternative(alternative)

Arguments

alternative

object of class DCAlternative or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAlternatives()

Set DC alternatives

Usage

DCEntry$setDCAlternatives(alternatives)

Arguments

alternatives

vector of class character

Method getDCAlternatives()

Get DC alternatives

Usage

DCEntry$getDCAlternatives()

Returns

a list of objects of class DCAlternative

Method addDCAudience()

Adds DC audience

Usage

DCEntry$addDCAudience(audience)

Arguments

audience

object of class DCAudience or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAudience()

Deletes DC audience

Usage

DCEntry$delDCAudience(audience)

Arguments

audience

object of class DCAudience or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAudiences()

Set DC audiences

Usage

DCEntry$setDCAudiences(audiences)

Arguments

audiences

vector of class character

Method getDCAudiences()

Get DC audiences

Usage

DCEntry$getDCAudiences()

Returns

a list of objects of class DCAudience

Method addDCAvailable()

Adds DC available

Usage

DCEntry$addDCAvailable(available)

Arguments

available

object of class DCAvailable or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAvailable()

Deletes DC available

Usage

DCEntry$delDCAvailable(available)

Arguments

available

object of class DCAvailable or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCAvailables()

Set DC availables

Usage

DCEntry$setDCAvailables(availables)

Arguments

availables

vector of class character

Method getDCAvailables()

Get DC availables

Usage

DCEntry$getDCAvailables()

Returns

a list of objects of class DCAvailable

Method addDCBibliographicCitation()

Adds DC bibliographic citation

Usage

DCEntry$addDCBibliographicCitation(bibliographicCitation)

Arguments

bibliographicCitation

object of class DCBibliographicCitation or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCBibliographicCitation()

Deletes DC bibliographic citation

Usage

DCEntry$delDCBibliographicCitation(bibliographicCitation)

Arguments

bibliographicCitation

object of class DCBibliographicCitation or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCBibliographicCitations()

Set bibliographic citations

Usage

DCEntry$setDCBibliographicCitations(bibliographicCitations)

Arguments

bibliographicCitations

vector of class character

Method getDCBibliographicCitations()

Get bibliographic citations

Usage

DCEntry$getDCBibliographicCitations()

Returns

the list of objects of class DCBibliographicCitation

Method addDCConformsTo()

Adds DC conforms to

Usage

DCEntry$addDCConformsTo(conformsTo)

Arguments

conformsTo

object of class DCConformsTo or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCConformsTo()

Deletes DC conforms to

Usage

DCEntry$delDCConformsTo(conformsTo)

Arguments

conformsTo

object of class DCConformsTo or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCConformsTo()

Set DC conforms to

Usage

DCEntry$setDCConformsTo(conformsTo)

Arguments

conformsTo

vector of class character

Method getDCConformsTo()

Get DC conforms to

Usage

DCEntry$getDCConformsTo()

Returns

the list of objects of class DCConformsTo

Method addDCContributor()

Adds DC contributor

Usage

DCEntry$addDCContributor(contributor)

Arguments

contributor

object of class DCContributor or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCContributor()

Deletes DC contributor

Usage

DCEntry$delDCContributor(contributor)

Arguments

contributor

object of class DCContributor or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCContributors()

Set DC contributors

Usage

DCEntry$setDCContributors(contributors)

Arguments

contributors

vector of class character

Method getDCContributors()

Get DC contributors

Usage

DCEntry$getDCContributors()

Returns

list of objects of class DCContributor

Method addDCCoverage()

Adds DC coverage

Usage

DCEntry$addDCCoverage(coverage)

Arguments

coverage

object of class DCCoverage or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCoverage()

Deletes DC coverage

Usage

DCEntry$delDCCoverage(coverage)

Arguments

coverage

object of class DCCoverage or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCCoverages()

Set DC coverages

Usage

DCEntry$setDCCoverages(coverages)

Arguments

coverages

coverages vector of class character

Method getDCCoverages()

Get DC coverages

Usage

DCEntry$getDCCoverages()

Returns

a list of objects of class DCCoverage

Method addDCCreated()

Adds DC created

Usage

DCEntry$addDCCreated(created)

Arguments

created

object of class DCCreated or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCreated()

Deletes DC created

Usage

DCEntry$delDCCreated(created)

Arguments

created

object of class DCCreated or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCCreator()

Adds DC creator

Usage

DCEntry$addDCCreator(creator)

Arguments

creator

object of class DCCreator or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCreator()

Deletes DC creator

Usage

DCEntry$delDCCreator(creator)

Arguments

creator

object of class DCCreator or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCCreators()

Set DC creators

Usage

DCEntry$setDCCreators(creators)

Arguments

creators

creators

Method getDCCreators()

Get DC creators

Usage

DCEntry$getDCCreators()

Returns

a list of objects of class DCCreator

Method addDCDate()

Adds DC date

Usage

DCEntry$addDCDate(date)

Arguments

date

object of class DCDate or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDate()

Deletes DC date

Usage

DCEntry$delDCDate(date)

Arguments

date

object of class DCDate or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCDates()

Set DC Creators

Usage

DCEntry$setDCDates(dates)

Arguments

dates

dates vector of class Date or POSIXt

Method getDCDates()

Get DC Dates

Usage

DCEntry$getDCDates()

Returns

a list of objects of class DCDate

Method addDCDateAccepted()

Adds DC date accepted

Usage

DCEntry$addDCDateAccepted(dateAccepted)

Arguments

dateAccepted

object of class DCDateAccepted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateAccepted()

Deletes DC date accepted

Usage

DCEntry$delDCDateAccepted(dateAccepted)

Arguments

dateAccepted

object of class DCDateAccepted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDateCopyrighted()

Adds DC date copyrighted

Usage

DCEntry$addDCDateCopyrighted(dateCopyrighted)

Arguments

dateCopyrighted

object of class DCDateCopyrighted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateCopyrighted()

Deletes DC date copyrighted

Usage

DCEntry$delDCDateCopyrighted(dateCopyrighted)

Arguments

dateCopyrighted

object of class DCDateCopyrighted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDateSubmitted()

Adds DC date submitted

Usage

DCEntry$addDCDateSubmitted(dateSubmitted)

Arguments

dateSubmitted

object of class DCDateSubmitted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateSubmitted()

Deletes DC date submitted

Usage

DCEntry$delDCDateSubmitted(dateSubmitted)

Arguments

dateSubmitted

object of class DCDateSubmitted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDescription()

Adds DC description

Usage

DCEntry$addDCDescription(description)

Arguments

description

object of class DCDescription or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDescription()

Deletes DC description

Usage

DCEntry$delDCDescription(description)

Arguments

description

object of class DCDescription or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCDescriptions()

Set DC descriptions

Usage

DCEntry$setDCDescriptions(descriptions)

Arguments

descriptions

vector of class character

Method getDCDescriptions()

Get DC descriptions

Usage

DCEntry$getDCDescriptions()

Returns

a list of objects of class DCDescription

Method addDCEducationalLevel()

Adds DC educational level

Usage

DCEntry$addDCEducationalLevel(educationalLevel)

Arguments

educationalLevel

object of class DCEducationalLevel or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCEducationalLevel()

Deletes DC educational level

Usage

DCEntry$delDCEducationalLevel(educationalLevel)

Arguments

educationalLevel

object of class DCEducationalLevel or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCEducationalLevels()

set DC education levels

Usage

DCEntry$setDCEducationalLevels(educationLevels)

Arguments

educationLevels

vector of class character

Method getDCEducationalLevels()

Get DC educational levels

Usage

DCEntry$getDCEducationalLevels()

Returns

a list of objects of class DCEducationalLevel

Method addDCExtent()

Adds DC extent

Usage

DCEntry$addDCExtent(extent)

Arguments

extent

object of class DCExtent or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCExtent()

Deletes DC extent

Usage

DCEntry$delDCExtent(extent)

Arguments

extent

object of class DCExtent or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCExtents()

Set DC extents

Usage

DCEntry$setDCExtents(extents)

Arguments

extents

vector of class character

Method getDCExtents()

Get DC extents

Usage

DCEntry$getDCExtents()

Returns

a list of objects of class DCExtent

Method addDCFormat()

Adds DC format

Usage

DCEntry$addDCFormat(format)

Arguments

format

object of class DCFormat or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCFormat()

Deletes DC format

Usage

DCEntry$delDCFormat(format)

Arguments

format

object of class DCFormat or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCFormats()

Set DC formats

Usage

DCEntry$setDCFormats(formats)

Arguments

formats

vector of class character

Method getDCFormats()

Get DC formats

Usage

DCEntry$getDCFormats()

Returns

a list of objects of class DCFormat

Method addDCHasPart()

Adds DC hasPart

Usage

DCEntry$addDCHasPart(hasPart)

Arguments

hasPart

object of class DCHasPart or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCHasPart()

Deletes DC hasPart

Usage

DCEntry$delDCHasPart(hasPart)

Arguments

hasPart

object of class DCHasPart or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCHasParts()

Set DC hasParts

Usage

DCEntry$setDCHasParts(hasParts)

Arguments

hasParts

vector of class character

Method getDCHasParts()

Get DC has part

Usage

DCEntry$getDCHasParts()

Returns

a list of objects of class DCHasPart

Method addDCHasVersion()

Adds DC hasVersion

Usage

DCEntry$addDCHasVersion(hasVersion)

Arguments

hasVersion

object of class DCHasVersion or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCHasVersion()

Deletes DC hasVersion

Usage

DCEntry$delDCHasVersion(hasVersion)

Arguments

hasVersion

object of class DCHasVersion or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCHasVersions()

Set DC hasVersions

Usage

DCEntry$setDCHasVersions(hasVersions)

Arguments

hasVersions

vector of class character

Method getDCHasVersions()

Get DC has versions

Usage

DCEntry$getDCHasVersions()

Returns

a list of objects of class DCHasVersion

Method addDCIdentifier()

Adds DC identifier

Usage

DCEntry$addDCIdentifier(identifier)

Arguments

identifier

object of class DCIdentifier or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIdentifier()

Deletes DC identifier

Usage

DCEntry$delDCIdentifier(identifier)

Arguments

identifier

object of class DCIdentifier or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIdentifiers()

Set DC identifiers

Usage

DCEntry$setDCIdentifiers(identifiers)

Arguments

identifiers

vector of class character

Method getDCIdentifiers()

Get DC identifiers

Usage

DCEntry$getDCIdentifiers()

Returns

a list of objects of class DCIdentifier

Method addDCInstructionalMethod()

Adds DC instructionalMethod

Usage

DCEntry$addDCInstructionalMethod(instructionalMethod)

Arguments

instructionalMethod

object of class DCInstructionalMethod or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCInstructionalMethod()

Deletes DC instructionalMethod

Usage

DCEntry$delDCInstructionalMethod(instructionalMethod)

Arguments

instructionalMethod

object of class DCInstructionalMethod or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCInstructionalMethods()

Set DC Instructional methods

Usage

DCEntry$setDCInstructionalMethods(instructionalMethods)

Arguments

instructionalMethods

vector of class character

Method getDCInstructionalMethods()

Get DC instructional methods

Usage

DCEntry$getDCInstructionalMethods()

Returns

a list of objects of class DCInstructionalMethod

Method addDCIsPartOf()

Adds DC isPartOf

Usage

DCEntry$addDCIsPartOf(isPartOf)

Arguments

isPartOf

object of class DCIsPartOf or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsPartOf()

Deletes DC isPartOf

Usage

DCEntry$delDCIsPartOf(isPartOf)

Arguments

isPartOf

object of class DCIsPartOf or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIsPartOf()

Set DC IsPartOf

Usage

DCEntry$setDCIsPartOf(isPartOf)

Arguments

isPartOf

vector of class character

Method getDCIsPartOfs()

Get DC Is Part of

Usage

DCEntry$getDCIsPartOfs()

Returns

a list of objects of class DCIsPartOf

Method addDCIsReferencedBy()

Adds DC isReferencedBy

Usage

DCEntry$addDCIsReferencedBy(isReferencedBy)

Arguments

isReferencedBy

object of class DCIsReferencedBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsReferencedBy()

Deletes DC isReferencedBy

Usage

DCEntry$delDCIsReferencedBy(isReferencedBy)

Arguments

isReferencedBy

object of class DCIsReferencedBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIsReferencedBys()

Set DC isReferencedBys

Usage

DCEntry$setDCIsReferencedBys(isReferencedBys)

Arguments

isReferencedBys

vector of class character

Method getDCIsReferencedBys()

Get DC Is Referenced by

Usage

DCEntry$getDCIsReferencedBys()

Returns

a list of objects of class DCIsReferencedBy

Method addDCIsReplacedBy()

Adds DC isReplacedBy

Usage

DCEntry$addDCIsReplacedBy(isReplacedBy)

Arguments

isReplacedBy

object of class DCIsReplacedBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsReplacedBy()

Deletes DC isReferencedBy

Usage

DCEntry$delDCIsReplacedBy(isReplacedBy)

Arguments

isReplacedBy

object of class DCIsReplacedBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIsReplacedBys()

Set DC isReplacedBys

Usage

DCEntry$setDCIsReplacedBys(isReplacedBys)

Arguments

isReplacedBys

vector of class character

Method getDCIsReplacedBys()

Get DC Is Replaced by

Usage

DCEntry$getDCIsReplacedBys()

Returns

a list of objects of class DCIsReplacedBy

Method addDCIsRequiredBy()

Adds DC isRequiredBy

Usage

DCEntry$addDCIsRequiredBy(isRequiredBy)

Arguments

isRequiredBy

object of class DCIsRequiredBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsRequiredBy()

Deletes DC isRequiredBy

Usage

DCEntry$delDCIsRequiredBy(isRequiredBy)

Arguments

isRequiredBy

object of class DCIsRequiredBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIsRequiredBys()

Set DC isRequiredBys

Usage

DCEntry$setDCIsRequiredBys(isRequiredBys)

Arguments

isRequiredBys

vector of class character

Method getDCIsRequiredBys()

Get DC Is Required by

Usage

DCEntry$getDCIsRequiredBys()

Returns

a list of objects of class DCIsRequiredBy

Method addDCIsVersionOf()

Adds DC isVersionOf

Usage

DCEntry$addDCIsVersionOf(isVersionOf)

Arguments

isVersionOf

object of class DCIsVersionOf or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsVersionOf()

Deletes DC isVersionOf

Usage

DCEntry$delDCIsVersionOf(isVersionOf)

Arguments

isVersionOf

object of class DCIsVersionOf or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCIsVersionOfs()

Set DC isVersionOfs

Usage

DCEntry$setDCIsVersionOfs(isVersionOfs)

Arguments

isVersionOfs

vector of class character

Method getDCIsVersionOfs()

Get DC Is Version Ofs

Usage

DCEntry$getDCIsVersionOfs()

Returns

a list of objects of class DCIsVersionOf

Method addDCIssued()

Adds DC issued

Usage

DCEntry$addDCIssued(issued)

Arguments

issued

object of class DCIssued or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIssued()

Deletes DC issued

Usage

DCEntry$delDCIssued(issued)

Arguments

issued

object of class DCIssued or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCLanguage()

Adds DC language

Usage

DCEntry$addDCLanguage(language)

Arguments

language

object of class DCLanguage or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCLanguage()

Deletes DC language

Usage

DCEntry$delDCLanguage(language)

Arguments

language

object of class DCLanguage or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCLanguages()

Set DC languages

Usage

DCEntry$setDCLanguages(languages)

Arguments

languages

languages vector of class character

Method getDCLanguages()

Get languages

Usage

DCEntry$getDCLanguages()

Returns

a list of objects of class DCLanguage

Method addDCLicense()

Adds DC license

Usage

DCEntry$addDCLicense(license)

Arguments

license

object of class DCLicense or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCLicense()

Deletes DC license

Usage

DCEntry$delDCLicense(license)

Arguments

license

object of class DCLicense or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCLicenses()

Set DC licences

Usage

DCEntry$setDCLicenses(licenses)

Arguments

licenses

vector of class character

Method getDCLicenses()

Get DC licenses

Usage

DCEntry$getDCLicenses()

Returns

a list of objects of class DCLicense

Method addDCMediator()

Adds DC mediator

Usage

DCEntry$addDCMediator(mediator)

Arguments

mediator

object of class DCMediator or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCMediator()

Deletes DC mediator

Usage

DCEntry$delDCMediator(mediator)

Arguments

mediator

object of class DCMediator or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCMediators()

Set DC mediators

Usage

DCEntry$setDCMediators(mediators)

Arguments

mediators

vector of class character

Method getDCMediators()

Get DC mediators

Usage

DCEntry$getDCMediators()

Returns

a list of objects of class DCMediator

Method addDCMedium()

Adds DC medium

Usage

DCEntry$addDCMedium(medium)

Arguments

medium

object of class DCMedium or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCMedium()

Deletes DC medium

Usage

DCEntry$delDCMedium(medium)

Arguments

medium

object of class DCMedium or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCMediums()

Set DC mediums

Usage

DCEntry$setDCMediums(mediums)

Arguments

mediums

vector of class character

Method getDCMediums()

Get DC mediums

Usage

DCEntry$getDCMediums()

Returns

a list of objects of class DCMedium

Method addDCModified()

Adds DC modified

Usage

DCEntry$addDCModified(modified)

Arguments

modified

object of class DCModified or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCModified()

Deletes DC modified

Usage

DCEntry$delDCModified(modified)

Arguments

modified

object of class DCModified or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deletes, FALSE otherwise

Method addDCProvenance()

Adds DC provenance

Usage

DCEntry$addDCProvenance(provenance)

Arguments

provenance

object of class DCProvenance or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCProvenance()

Deletes DC provenance

Usage

DCEntry$delDCProvenance(provenance)

Arguments

provenance

object of class DCProvenance or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCProvenances()

Set DC provenances

Usage

DCEntry$setDCProvenances(provenances)

Arguments

provenances

vector of class character

Method getDCProvenances()

Get DC provenances

Usage

DCEntry$getDCProvenances()

Returns

a list of objects of class DCProvenance

Method addDCPublisher()

Adds DC publisher

Usage

DCEntry$addDCPublisher(publisher)

Arguments

publisher

object of class DCPublisher or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCPublisher()

Deletes DC publisher

Usage

DCEntry$delDCPublisher(publisher)

Arguments

publisher

object of class DCPublisher or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCPublishers()

Set DC publishers

Usage

DCEntry$setDCPublishers(publishers)

Arguments

publishers

vector of class character

Method getDCPublishers()

Get DC publishers

Usage

DCEntry$getDCPublishers()

Returns

a list of objects of class DCPublisher

Method addDCReferences()

Adds DC references

Usage

DCEntry$addDCReferences(references)

Arguments

references

object of class DCReferences or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCReferences()

Deletes DC references

Usage

DCEntry$delDCReferences(references)

Arguments

references

object of class DCReferences or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCReferences()

Set DC references

Usage

DCEntry$setDCReferences(references)

Arguments

references

vector of class character

Method getDCReferences()

Get DC references

Usage

DCEntry$getDCReferences()

Returns

a list of objects of class DCReferences

Method addDCRelation()

Adds DC relation

Usage

DCEntry$addDCRelation(relation)

Arguments

relation

object of class DCRelation or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRelation()

Deletes DC relation

Usage

DCEntry$delDCRelation(relation)

Arguments

relation

object of class DCRelation or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCRelations()

Set DC relations

Usage

DCEntry$setDCRelations(relations)

Arguments

relations

vector of class character

Method getDCRelations()

Get DC relations

Usage

DCEntry$getDCRelations()

Returns

a list of objects of class DCRelation

Method addDCReplaces()

Adds DC replaces

Usage

DCEntry$addDCReplaces(replaces)

Arguments

replaces

object of class DCReplaces or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCReplaces()

Deletes DC replaces

Usage

DCEntry$delDCReplaces(replaces)

Arguments

replaces

object of class DCReplaces or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCReplaces()

Set DC replaces

Usage

DCEntry$setDCReplaces(replaces)

Arguments

replaces

vector of class character

Method getDCReplaces()

Get DC replaces

Usage

DCEntry$getDCReplaces()

Returns

a list of objects of class DCReplaces

Method addDCRequires()

Adds DC requires

Usage

DCEntry$addDCRequires(requires)

Arguments

requires

object of class DCRequires or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRequires()

Deletes DC requires

Usage

DCEntry$delDCRequires(requires)

Arguments

requires

object of class DCRequires or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCRequires()

Set DC requires

Usage

DCEntry$setDCRequires(requires)

Arguments

requires

vector of class character

Method getDCRequires()

Get DC requires

Usage

DCEntry$getDCRequires()

Returns

a list of objects of class DCRequires

Method addDCRights()

Adds DC rights

Usage

DCEntry$addDCRights(rights)

Arguments

rights

object of class DCRights or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRights()

Deletes DC rights

Usage

DCEntry$delDCRights(rights)

Arguments

rights

object of class DCRights or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCRights()

Set DC rights

Usage

DCEntry$setDCRights(rights)

Arguments

rights

vector of class character

Method getDCRights()

Get DC rights

Usage

DCEntry$getDCRights()

Returns

a list of objects of class DCRights

Method addDCRightsHolder()

Adds DC rightsHolder

Usage

DCEntry$addDCRightsHolder(rightsHolder)

Arguments

rightsHolder

object of class DCRightsHolder or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRightsHolder()

Deletes DC rightsHolder

Usage

DCEntry$delDCRightsHolder(rightsHolder)

Arguments

rightsHolder

object of class DCRightsHolder or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCRightsHolders()

Set DC rights holders

Usage

DCEntry$setDCRightsHolders(rightsHolders)

Arguments

rightsHolders

vector of class character

Method getDCRightsHolders()

Get DC rights holders

Usage

DCEntry$getDCRightsHolders()

Returns

a list of objects of class DCRightsHolder

Method addDCSource()

Adds DC source

Usage

DCEntry$addDCSource(source)

Arguments

source

object of class DCSource or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCSource()

Deletes DC source

Usage

DCEntry$delDCSource(source)

Arguments

source

object of class DCSource or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCSources()

Set DC sources

Usage

DCEntry$setDCSources(sources)

Arguments

sources

vector of class character

Method getDCSources()

Get DC sources

Usage

DCEntry$getDCSources()

Returns

a list of objects of class DCSource

Method addDCSubject()

Adds DC subject

Usage

DCEntry$addDCSubject(subject)

Arguments

subject

object of class DCSubject or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCSubject()

Deletes DC subject

Usage

DCEntry$delDCSubject(subject)

Arguments

subject

object of class DCSubject or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCSubjects()

Set DC subjects

Usage

DCEntry$setDCSubjects(subjects)

Arguments

subjects

vector of class character

Method getDCSubjects()

Get DC Subjects

Usage

DCEntry$getDCSubjects()

Returns

a list of objects of class DCSubject

Method addDCTableOfContents()

Adds DC tableOfContents

Usage

DCEntry$addDCTableOfContents(tableOfContents)

Arguments

tableOfContents

object of class DCTableOfContents or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTableOfContents()

Deletes DC tableOfContents

Usage

DCEntry$delDCTableOfContents(tableOfContents)

Arguments

tableOfContents

object of class DCTableOfContents or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCTablesOfContents()

Set DC tables of contents

Usage

DCEntry$setDCTablesOfContents(tablesOfContents)

Arguments

tablesOfContents

vector of class character

Method getDCTablesOfContent()

Get DC tables of contents

Usage

DCEntry$getDCTablesOfContent()

Returns

a list of objects of class DCTableOfContents

Method addDCTemporal()

Adds DC temporal

Usage

DCEntry$addDCTemporal(temporal)

Arguments

temporal

object of class DCTemporal or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTemporal()

Deletes DC temporal

Usage

DCEntry$delDCTemporal(temporal)

Arguments

temporal

object of class DCTemporal or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCTemporals()

Set DC temporals

Usage

DCEntry$setDCTemporals(temporals)

Arguments

temporals

vector of class character

Method getDCTemporals()

Get DC temporals

Usage

DCEntry$getDCTemporals()

Returns

a list of objects of class DCTemporal

Method addDCTitle()

Adds DC title

Usage

DCEntry$addDCTitle(title)

Arguments

title

object of class DCTitle or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTitle()

Deletes DC title

Usage

DCEntry$delDCTitle(title)

Arguments

title

object of class DCTitle or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCTitles()

Set DC titles

Usage

DCEntry$setDCTitles(titles)

Arguments

titles

vector of class character

Method getDCTitles()

Get DC titles

Usage

DCEntry$getDCTitles()

Returns

a list of objects of class DCTitle

Method addDCType()

Adds DC type

Usage

DCEntry$addDCType(type)

Arguments

type

object of class DCType or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCType()

Deletes DC type

Usage

DCEntry$delDCType(type)

Arguments

type

object of class DCType or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method setDCTypes()

Set DC Types

Usage

DCEntry$setDCTypes(types)

Arguments

types

vector of class character

Method getDCTypes()

Get DC types

Usage

DCEntry$getDCTypes()

Returns

a list of objects of class DCType

Method asDataFrame()

export to a data.frame

Usage

DCEntry$asDataFrame()

Returns

an object of class data.frame

Method clone()

The objects of this class are cloneable with this method.

Usage

DCEntry$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 # \donttest{
  #encoding
  dcentry <- DCEntry$new()
  dcentry$setId("my-dc-entry")

  #fill dc entry
  dcentry$addDCDate(Sys.time())
  dcentry$addDCTitle("atom4R - Tools to read/write and publish metadata as Atom XML format")
  dcentry$addDCType("Software")
  creator <- DCCreator$new(value = "Blondel, Emmanuel")
  creator$attrs[["affiliation"]] <- "Independent"
  dcentry$addDCCreator(creator)
  dcentry$addDCSubject("R")
  dcentry$addDCSubject("FAIR")
  dcentry$addDCSubject("Interoperability")
  dcentry$addDCSubject("Open Science")
  dcentry$addDCDescription("Atom4R offers tools to read/write and publish metadata as Atom XML")
  dcentry$addDCPublisher("GitHub")
  funder <- DCContributor$new(value = "CNRS")
  funder$attrs[["type"]] <- "Funder"
  dcentry$addDCContributor(funder)
  dcentry$addDCRelation("Github repository: https://github.com/eblondel/atom4R")
  dcentry$addDCSource("Atom Syndication format - https://www.ietf.org/rfc/rfc4287")
  dcentry$addDCSource("AtomPub, The Atom publishing protocol - https://tools.ietf.org/html/rfc5023")
  dcentry$addDCSource("Sword API - http://swordapp.org/")
  dcentry$addDCSource("Dublin Core Metadata Initiative - https://www.dublincore.org/")
  dcentry$addDCSource("Guidelines for implementing Dublin Core in XML")
  dcentry$addDCLicense("NONE")
  dcentry$addDCRights("MIT License")
  dcentry$addDCHasPart("part1")
  dcentry$addDCHasPart("part2")
  dcentry$addDCHasVersion("0.2")
  dcentry$addDCIsPartOf("CRAN")
  dcentry$addDCIsPartOf("GitHub")
  dcentry$addDCIsReferencedBy("CRAN")
  dcentry$addDCIsReferencedBy("GitHub")
  dcentry$addDCIsRequiredBy("zen4R")
  dcentry$addDCIsRequiredBy("cloud4R")

  xml <- dcentry$encode()
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}contributor', attribute 'type': The attribute 'type' is not allowed at line 11. 
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}creator', attribute 'affiliation': The attribute 'affiliation' is not allowed at line 14. 
#> [atom4R][WARN] Object 'DCEntry' is INVALID according to Atom XML schemas! 

  #decoding
  dcentry2 <- DCEntry$new(xml = xml)
  xml2 <- dcentry2$encode()
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}contributor', attribute 'type': The attribute 'type' is not allowed at line 11. 
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}creator', attribute 'affiliation': The attribute 'affiliation' is not allowed at line 14. 
#> [atom4R][WARN] Object 'DCEntry' is INVALID according to Atom XML schemas! 
 # }