SwordHalClient class

This class models an Sword service client for HAL (Archives Houvertes)

Format

R6Class object.

Value

Object of R6Class for modelling an Sword client

Details

SwordHalClient

Note

Experimental

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

atom4R::atom4RLogger -> atom4R::AtomPubClient -> atom4R::SwordClient -> SwordHalClient

Methods

Public methods

SwordHalClient$new()

SwordHalClient$getServiceDocument()

SwordHalClient$getCollectionMembers()

SwordHalClient$clone()

Inherited methods

atom4R::atom4RLogger$ERROR()

atom4R::atom4RLogger$INFO()

atom4R::atom4RLogger$WARN()

atom4R::atom4RLogger$getClass()

atom4R::atom4RLogger$getClassName()

atom4R::atom4RLogger$logger()

atom4R::AtomPubClient$getPwd()

atom4R::AtomPubClient$getToken()

atom4R::AtomPubClient$getUser()

atom4R::AtomPubClient$listCollections()

Method new()

This method is to instantiate an Sword HAL (Archive Ouvertes - https://hal.archives-ouvertes.fr/) Client. By default the version is set to "2".

The keyring_backend can be set to use a different backend for storing the SWORD API user token with keyring (Default value is 'env').

The logger allows to specify the level of log (default is NULL), either "INFO" for atom4R logs or "DEBUG" for verbose HTTP client (curl) logs.

Usage

SwordHalClient$new(
  url,
  user = NULL,
  pwd = NULL,
  logger = NULL,
  keyring_backend = "env"
)

Arguments

url

url

user

user

pwd

pwd

logger

logger

keyring_backend

keyring backend. Default value is 'env'

Method getServiceDocument()

Get service document

Usage

SwordHalClient$getServiceDocument(force = FALSE)

Arguments

force

force Force getting/refreshing of service document

Returns

object of class SwordServiceDocument

Method getCollectionMembers()

Get collection members

Usage

SwordHalClient$getCollectionMembers(collectionId)

Arguments

collectionId

collection ID

Returns

a list of AtomFeed

Method clone()

The objects of this class are cloneable with this method.

Usage

SwordHalClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.