Dublin Core Entry class

This class models an Dublin Core Entry

Format

R6Class object.

Value

Object of R6Class for modelling an Dublin Core Entry

Details

DCEntry

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

atom4R::atom4RLogger -> atom4R::AtomAbstractObject -> atom4R::AtomEntry -> DCEntry

Methods

Public methods

DCEntry$new()

DCEntry$addDCElement()

DCEntry$delDCElement()

DCEntry$addDCAbstract()

DCEntry$delDCAbstract()

DCEntry$addDCAccessRights()

DCEntry$delDCAccessRights()

DCEntry$addDCAccrualMethod()

DCEntry$delDCAccrualMethod()

DCEntry$addDCAccrualPeriodicity()

DCEntry$delDCAccrualPeriodicity()

DCEntry$addDCAccrualPolicy()

DCEntry$delDCAccrualPolicy()

DCEntry$addDCAlternative()

DCEntry$delDCAlternative()

DCEntry$addDCAudience()

DCEntry$delDCAudience()

DCEntry$addDCAvailable()

DCEntry$delDCAvailable()

DCEntry$addDCBibliographicCitation()

DCEntry$delDCBibliographicCitation()

DCEntry$addDCConformsTo()

DCEntry$delDCConformsTo()

DCEntry$addDCContributor()

DCEntry$delDCContributor()

DCEntry$addDCCoverage()

DCEntry$delDCCoverage()

DCEntry$addDCCreated()

DCEntry$delDCCreated()

DCEntry$addDCCreator()

DCEntry$delDCCreator()

DCEntry$addDCDate()

DCEntry$delDCDate()

DCEntry$addDCDateAccepted()

DCEntry$delDCDateAccepted()

DCEntry$addDCDateCopyrighted()

DCEntry$delDCDateCopyrighted()

DCEntry$addDCDateSubmitted()

DCEntry$delDCDateSubmitted()

DCEntry$addDCDescription()

DCEntry$delDCDescription()

DCEntry$addDCEducationalLevel()

DCEntry$delDCEducationalLevel()

DCEntry$addDCExtent()

DCEntry$delDCExtent()

DCEntry$addDCFormat()

DCEntry$delDCFormat()

DCEntry$addDCHasPart()

DCEntry$delDCHasPart()

DCEntry$addDCHasVersion()

DCEntry$delDCHasVersion()

DCEntry$addDCIdentifier()

DCEntry$delDCIdentifier()

DCEntry$addDCInstructionalMethod()

DCEntry$delDCInstructionalMethod()

DCEntry$addDCIsPartOf()

DCEntry$delDCIsPartOf()

DCEntry$addDCIsReferencedBy()

DCEntry$delDCIsReferencedBy()

DCEntry$addDCIsReplacedBy()

DCEntry$delDCIsReplacedBy()

DCEntry$addDCIsRequiredBy()

DCEntry$delDCIsRequiredBy()

DCEntry$addDCIsVersionOf()

DCEntry$delDCIsVersionOf()

DCEntry$addDCIssued()

DCEntry$delDCIssued()

DCEntry$addDCLanguage()

DCEntry$delDCLanguage()

DCEntry$addDCLicense()

DCEntry$delDCLicense()

DCEntry$addDCMediator()

DCEntry$delDCMediator()

DCEntry$addDCMedium()

DCEntry$delDCMedium()

DCEntry$addDCModified()

DCEntry$delDCModified()

DCEntry$addDCProvenance()

DCEntry$delDCProvenance()

DCEntry$addDCPublisher()

DCEntry$delDCPublisher()

DCEntry$addDCReferences()

DCEntry$delDCReferences()

DCEntry$addDCRelation()

DCEntry$delDCRelation()

DCEntry$addDCReplaces()

DCEntry$delDCReplaces()

DCEntry$addDCRequires()

DCEntry$delDCRequires()

DCEntry$addDCRights()

DCEntry$delDCRights()

DCEntry$addDCRightsHolder()

DCEntry$delDCRightsHolder()

DCEntry$addDCSource()

DCEntry$delDCSource()

DCEntry$addDCSubject()

DCEntry$delDCSubject()

DCEntry$addDCTableOfContents()

DCEntry$delDCTableOfContents()

DCEntry$addDCTemporal()

DCEntry$delDCTemporal()

DCEntry$addDCTitle()

DCEntry$delDCTitle()

DCEntry$addDCType()

DCEntry$delDCType()

DCEntry$clone()

Inherited methods

atom4R::atom4RLogger$ERROR()

atom4R::atom4RLogger$INFO()

atom4R::atom4RLogger$WARN()

atom4R::atom4RLogger$logger()

atom4R::AtomAbstractObject$addListElement()

atom4R::AtomAbstractObject$contains()

atom4R::AtomAbstractObject$createElement()

atom4R::AtomAbstractObject$decode()

atom4R::AtomAbstractObject$delListElement()

atom4R::AtomAbstractObject$encode()

atom4R::AtomAbstractObject$getClass()

atom4R::AtomAbstractObject$getClassName()

atom4R::AtomAbstractObject$getNamespace()

atom4R::AtomAbstractObject$getNamespaceDefinition()

atom4R::AtomAbstractObject$getRootElement()

atom4R::AtomAbstractObject$getXmlElement()

atom4R::AtomAbstractObject$isDocument()

atom4R::AtomAbstractObject$isFieldInheritedFrom()

atom4R::AtomAbstractObject$print()

atom4R::AtomAbstractObject$save()

atom4R::AtomAbstractObject$setIsDocument()

atom4R::AtomAbstractObject$validate()

atom4R::AtomEntry$addAuthor()

atom4R::AtomEntry$addCategory()

atom4R::AtomEntry$addContributor()

atom4R::AtomEntry$addLink()

atom4R::AtomEntry$delAuthor()

atom4R::AtomEntry$delCategory()

atom4R::AtomEntry$delContributor()

atom4R::AtomEntry$delLink()

atom4R::AtomEntry$setContent()

atom4R::AtomEntry$setId()

atom4R::AtomEntry$setPublished()

atom4R::AtomEntry$setRights()

atom4R::AtomEntry$setSource()

atom4R::AtomEntry$setSummary()

atom4R::AtomEntry$setTitle()

atom4R::AtomEntry$setUpdated()

Method new()

Initializes an object of class DCEntry

Usage

DCEntry$new(xml = NULL)

Arguments

xml

object of class XMLInternalNode-class from XML

Method addDCElement()

Adds a Dublin Core element

Usage

DCEntry$addDCElement(term, value)

Arguments

term

term

value

value

Returns

TRUE if added, FALSE otherwise

Method delDCElement()

Deletes a Dublin Core element

Usage

DCEntry$delDCElement(term, value)

Arguments

term

term

value

value

Returns

TRUE if deleted, FALSE otherwise

Method addDCAbstract()

Adds DC abstract

Usage

DCEntry$addDCAbstract(abstract)

Arguments

abstract

object of class DCAbstract or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAbstract()

Deletes DC abstract

Usage

DCEntry$delDCAbstract(abstract)

Arguments

abstract

object of class DCAbstract or vector of class character and length 1

Method addDCAccessRights()

Adds DC access rights

Usage

DCEntry$addDCAccessRights(accessRights)

Arguments

accessRights

object of class DCAccessRights or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccessRights()

Deletes DC access rights

Usage

DCEntry$delDCAccessRights(accessRights)

Arguments

accessRights

object of class DCAccessRights or vector of class character and length 1

Method addDCAccrualMethod()

Adds DC accrual method

Usage

DCEntry$addDCAccrualMethod(accrualMethod)

Arguments

accrualMethod

object of class DCAccrualMethod or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualMethod()

Deletes DC accrual method

Usage

DCEntry$delDCAccrualMethod(accrualMethod)

Arguments

accrualMethod

object of class DCAccrualMethod or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCAccrualPeriodicity()

Adds DC accrual periodicity

Usage

DCEntry$addDCAccrualPeriodicity(accrualPeriodicity)

Arguments

accrualPeriodicity

object of class DCAccrualPeriodicity or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualPeriodicity()

Deletes DC accrual periodicity

Usage

DCEntry$delDCAccrualPeriodicity(accrualPeriodicity)

Arguments

accrualPeriodicity

object of class DCAccrualPeriodicity or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCAccrualPolicy()

Adds DC accrual policy

Usage

DCEntry$addDCAccrualPolicy(accrualPolicy)

Arguments

accrualPolicy

object of class DCAccrualPolicy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAccrualPolicy()

Deletes DC accrual policy

Usage

DCEntry$delDCAccrualPolicy(accrualPolicy)

Arguments

accrualPolicy

object of class DCAccrualPolicy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCAlternative()

Adds DC alternative

Usage

DCEntry$addDCAlternative(alternative)

Arguments

alternative

object of class DCAlternative or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAlternative()

Deletes DC alternative

Usage

DCEntry$delDCAlternative(alternative)

Arguments

alternative

object of class DCAlternative or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCAudience()

Adds DC audience

Usage

DCEntry$addDCAudience(audience)

Arguments

audience

object of class DCAudience or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAudience()

Deletes DC audience

Usage

DCEntry$delDCAudience(audience)

Arguments

audience

object of class DCAudience or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCAvailable()

Adds DC available

Usage

DCEntry$addDCAvailable(available)

Arguments

available

object of class DCAvailable or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCAvailable()

Deletes DC available

Usage

DCEntry$delDCAvailable(available)

Arguments

available

object of class DCAvailable or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCBibliographicCitation()

Adds DC bibliographic citation

Usage

DCEntry$addDCBibliographicCitation(bibliographicCitation)

Arguments

bibliographicCitation

object of class DCBibliographicCitation or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCBibliographicCitation()

Deletes DC bibliographic citation

Usage

DCEntry$delDCBibliographicCitation(bibliographicCitation)

Arguments

bibliographicCitation

object of class DCBibliographicCitation or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCConformsTo()

Adds DC conforms to

Usage

DCEntry$addDCConformsTo(conformsTo)

Arguments

conformsTo

object of class DCConformsTo or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCConformsTo()

Deletes DC conforms to

Usage

DCEntry$delDCConformsTo(conformsTo)

Arguments

conformsTo

object of class DCConformsTo or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCContributor()

Adds DC contributor

Usage

DCEntry$addDCContributor(contributor)

Arguments

contributor

object of class DCContributor or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCContributor()

Deletes DC contributor

Usage

DCEntry$delDCContributor(contributor)

Arguments

contributor

object of class DCContributor or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCCoverage()

Adds DC coverage

Usage

DCEntry$addDCCoverage(coverage)

Arguments

coverage

object of class DCCoverage or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCoverage()

Deletes DC coverage

Usage

DCEntry$delDCCoverage(coverage)

Arguments

coverage

object of class DCCoverage or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCCreated()

Adds DC created

Usage

DCEntry$addDCCreated(created)

Arguments

created

object of class DCCreated or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCreated()

Deletes DC created

Usage

DCEntry$delDCCreated(created)

Arguments

created

object of class DCCreated or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCCreator()

Adds DC creator

Usage

DCEntry$addDCCreator(creator)

Arguments

creator

object of class DCCreator or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCCreator()

Deletes DC creator

Usage

DCEntry$delDCCreator(creator)

Arguments

creator

object of class DCCreator or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDate()

Adds DC date

Usage

DCEntry$addDCDate(date)

Arguments

date

object of class DCDate or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDate()

Deletes DC date

Usage

DCEntry$delDCDate(date)

Arguments

date

object of class DCDate or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDateAccepted()

Adds DC date accepted

Usage

DCEntry$addDCDateAccepted(dateAccepted)

Arguments

dateAccepted

object of class DCDateAccepted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateAccepted()

Deletes DC date accepted

Usage

DCEntry$delDCDateAccepted(dateAccepted)

Arguments

dateAccepted

object of class DCDateAccepted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDateCopyrighted()

Adds DC date copyrighted

Usage

DCEntry$addDCDateCopyrighted(dateCopyrighted)

Arguments

dateCopyrighted

object of class DCDateCopyrighted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateCopyrighted()

Deletes DC date copyrighted

Usage

DCEntry$delDCDateCopyrighted(dateCopyrighted)

Arguments

dateCopyrighted

object of class DCDateCopyrighted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDateSubmitted()

Adds DC date submitted

Usage

DCEntry$addDCDateSubmitted(dateSubmitted)

Arguments

dateSubmitted

object of class DCDateSubmitted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDateSubmitted()

Deletes DC date submitted

Usage

DCEntry$delDCDateSubmitted(dateSubmitted)

Arguments

dateSubmitted

object of class DCDateSubmitted or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCDescription()

Adds DC description

Usage

DCEntry$addDCDescription(description)

Arguments

description

object of class DCDescription or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCDescription()

Deletes DC description

Usage

DCEntry$delDCDescription(description)

Arguments

description

object of class DCDescription or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCEducationalLevel()

Adds DC educational level

Usage

DCEntry$addDCEducationalLevel(educationalLevel)

Arguments

educationalLevel

object of class DCEducationalLevel or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCEducationalLevel()

Deletes DC educational level

Usage

DCEntry$delDCEducationalLevel(educationalLevel)

Arguments

educationalLevel

object of class DCEducationalLevel or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCExtent()

Adds DC extent

Usage

DCEntry$addDCExtent(extent)

Arguments

extent

object of class DCExtent or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCExtent()

Deletes DC extent

Usage

DCEntry$delDCExtent(extent)

Arguments

extent

object of class DCExtent or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCFormat()

Adds DC format

Usage

DCEntry$addDCFormat(format)

Arguments

format

object of class DCFormat or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCFormat()

Deletes DC format

Usage

DCEntry$delDCFormat(format)

Arguments

format

object of class DCFormat or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCHasPart()

Adds DC hasPart

Usage

DCEntry$addDCHasPart(hasPart)

Arguments

hasPart

object of class DCHasPart or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCHasPart()

Deletes DC hasPart

Usage

DCEntry$delDCHasPart(hasPart)

Arguments

hasPart

object of class DCHasPart or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCHasVersion()

Adds DC hasVersion

Usage

DCEntry$addDCHasVersion(hasVersion)

Arguments

hasVersion

object of class DCHasVersion or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCHasVersion()

Deletes DC hasVersion

Usage

DCEntry$delDCHasVersion(hasVersion)

Arguments

hasVersion

object of class DCHasVersion or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIdentifier()

Adds DC identifier

Usage

DCEntry$addDCIdentifier(identifier)

Arguments

identifier

object of class DCIdentifier or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIdentifier()

Deletes DC identifier

Usage

DCEntry$delDCIdentifier(identifier)

Arguments

identifier

object of class DCIdentifier or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCInstructionalMethod()

Adds DC instructionalMethod

Usage

DCEntry$addDCInstructionalMethod(instructionalMethod)

Arguments

instructionalMethod

object of class DCInstructionalMethod or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCInstructionalMethod()

Deletes DC instructionalMethod

Usage

DCEntry$delDCInstructionalMethod(instructionalMethod)

Arguments

instructionalMethod

object of class DCInstructionalMethod or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIsPartOf()

Adds DC isPartOf

Usage

DCEntry$addDCIsPartOf(isPartOf)

Arguments

isPartOf

object of class DCIsPartOf or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsPartOf()

Deletes DC isPartOf

Usage

DCEntry$delDCIsPartOf(isPartOf)

Arguments

isPartOf

object of class DCIsPartOf or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIsReferencedBy()

Adds DC isReferencedBy

Usage

DCEntry$addDCIsReferencedBy(isReferencedBy)

Arguments

isReferencedBy

object of class DCIsReferencedBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsReferencedBy()

Deletes DC isReferencedBy

Usage

DCEntry$delDCIsReferencedBy(isReferencedBy)

Arguments

isReferencedBy

object of class DCIsReferencedBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIsReplacedBy()

Adds DC isReplacedBy

Usage

DCEntry$addDCIsReplacedBy(isReplacedBy)

Arguments

isReplacedBy

object of class DCIsReplacedBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsReplacedBy()

Deletes DC isReferencedBy

Usage

DCEntry$delDCIsReplacedBy(isReplacedBy)

Arguments

isReplacedBy

object of class DCIsReplacedBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIsRequiredBy()

Adds DC isRequiredBy

Usage

DCEntry$addDCIsRequiredBy(isRequiredBy)

Arguments

isRequiredBy

object of class DCIsRequiredBy or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsRequiredBy()

Deletes DC isRequiredBy

Usage

DCEntry$delDCIsRequiredBy(isRequiredBy)

Arguments

isRequiredBy

object of class DCIsRequiredBy or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIsVersionOf()

Adds DC isVersionOf

Usage

DCEntry$addDCIsVersionOf(isVersionOf)

Arguments

isVersionOf

object of class DCIsVersionOf or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIsVersionOf()

Deletes DC isVersionOf

Usage

DCEntry$delDCIsVersionOf(isVersionOf)

Arguments

isVersionOf

object of class DCIsVersionOf or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCIssued()

Adds DC issued

Usage

DCEntry$addDCIssued(issued)

Arguments

issued

object of class DCIssued or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCIssued()

Deletes DC issued

Usage

DCEntry$delDCIssued(issued)

Arguments

issued

object of class DCIssued or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCLanguage()

Adds DC language

Usage

DCEntry$addDCLanguage(language)

Arguments

language

object of class DCLanguage or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCLanguage()

Deletes DC language

Usage

DCEntry$delDCLanguage(language)

Arguments

language

object of class DCLanguage or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCLicense()

Adds DC license

Usage

DCEntry$addDCLicense(license)

Arguments

license

object of class DCLicense or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCLicense()

Deletes DC license

Usage

DCEntry$delDCLicense(license)

Arguments

license

object of class DCLicense or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCMediator()

Adds DC mediator

Usage

DCEntry$addDCMediator(mediator)

Arguments

mediator

object of class DCMediator or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCMediator()

Deletes DC mediator

Usage

DCEntry$delDCMediator(mediator)

Arguments

mediator

object of class DCMediator or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCMedium()

Adds DC medium

Usage

DCEntry$addDCMedium(medium)

Arguments

medium

object of class DCMedium or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCMedium()

Deletes DC medium

Usage

DCEntry$delDCMedium(medium)

Arguments

medium

object of class DCMedium or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCModified()

Adds DC modified

Usage

DCEntry$addDCModified(modified)

Arguments

modified

object of class DCModified or vector of class Date,POSIXt or character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCModified()

Deletes DC modified

Usage

DCEntry$delDCModified(modified)

Arguments

modified

object of class DCModified or vector of class Date,POSIXt or character and length 1

Returns

TRUE if deletes, FALSE otherwise

Method addDCProvenance()

Adds DC provenance

Usage

DCEntry$addDCProvenance(provenance)

Arguments

provenance

object of class DCProvenance or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCProvenance()

Deletes DC provenance

Usage

DCEntry$delDCProvenance(provenance)

Arguments

provenance

object of class DCProvenance or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCPublisher()

Adds DC publisher

Usage

DCEntry$addDCPublisher(publisher)

Arguments

publisher

object of class DCPublisher or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCPublisher()

Deletes DC publisher

Usage

DCEntry$delDCPublisher(publisher)

Arguments

publisher

object of class DCPublisher or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCReferences()

Adds DC references

Usage

DCEntry$addDCReferences(references)

Arguments

references

object of class DCReferences or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCReferences()

Deletes DC references

Usage

DCEntry$delDCReferences(references)

Arguments

references

object of class DCReferences or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCRelation()

Adds DC relation

Usage

DCEntry$addDCRelation(relation)

Arguments

relation

object of class DCRelation or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRelation()

Deletes DC relation

Usage

DCEntry$delDCRelation(relation)

Arguments

relation

object of class DCRelation or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCReplaces()

Adds DC replaces

Usage

DCEntry$addDCReplaces(replaces)

Arguments

replaces

object of class DCReplaces or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCReplaces()

Deletes DC replaces

Usage

DCEntry$delDCReplaces(replaces)

Arguments

replaces

object of class DCReplaces or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCRequires()

Adds DC requires

Usage

DCEntry$addDCRequires(requires)

Arguments

requires

object of class DCRequires or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRequires()

Deletes DC requires

Usage

DCEntry$delDCRequires(requires)

Arguments

requires

object of class DCRequires or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCRights()

Adds DC rights

Usage

DCEntry$addDCRights(rights)

Arguments

rights

object of class DCRights or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRights()

Deletes DC rights

Usage

DCEntry$delDCRights(rights)

Arguments

rights

object of class DCRights or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCRightsHolder()

Adds DC rightsHolder

Usage

DCEntry$addDCRightsHolder(rightsHolder)

Arguments

rightsHolder

object of class DCRightsHolder or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCRightsHolder()

Deletes DC rightsHolder

Usage

DCEntry$delDCRightsHolder(rightsHolder)

Arguments

rightsHolder

object of class DCRightsHolder or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCSource()

Adds DC source

Usage

DCEntry$addDCSource(source)

Arguments

source

object of class DCSource or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCSource()

Deletes DC source

Usage

DCEntry$delDCSource(source)

Arguments

source

object of class DCSource or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCSubject()

Adds DC subject

Usage

DCEntry$addDCSubject(subject)

Arguments

subject

object of class DCSubject or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCSubject()

Deletes DC subject

Usage

DCEntry$delDCSubject(subject)

Arguments

subject

object of class DCSubject or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCTableOfContents()

Adds DC tableOfContents

Usage

DCEntry$addDCTableOfContents(tableOfContents)

Arguments

tableOfContents

object of class DCTableOfContents or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTableOfContents()

Deletes DC tableOfContents

Usage

DCEntry$delDCTableOfContents(tableOfContents)

Arguments

tableOfContents

object of class DCTableOfContents or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCTemporal()

Adds DC temporal

Usage

DCEntry$addDCTemporal(temporal)

Arguments

temporal

object of class DCTemporal or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTemporal()

Deletes DC temporal

Usage

DCEntry$delDCTemporal(temporal)

Arguments

temporal

object of class DCTemporal or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCTitle()

Adds DC title

Usage

DCEntry$addDCTitle(title)

Arguments

title

object of class DCTitle or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCTitle()

Deletes DC title

Usage

DCEntry$delDCTitle(title)

Arguments

title

object of class DCTitle or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method addDCType()

Adds DC type

Usage

DCEntry$addDCType(type)

Arguments

type

object of class DCType or vector of class character and length 1

Returns

TRUE if added, FALSE otherwise

Method delDCType()

Deletes DC type

Usage

DCEntry$delDCType(type)

Arguments

type

object of class DCType or vector of class character and length 1

Returns

TRUE if deleted, FALSE otherwise

Method clone()

The objects of this class are cloneable with this method.

Usage

DCEntry$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

 # \donttest{
  #encoding
  dcentry <- DCEntry$new()
  dcentry$setId("my-dc-entry")

  #fill dc entry
  dcentry$addDCDate(Sys.time())
  dcentry$addDCTitle("atom4R - Tools to read/write and publish metadata as Atom XML format")
  dcentry$addDCType("Software")
  creator <- DCCreator$new(value = "Blondel, Emmanuel")
  creator$attrs[["affiliation"]] <- "Independent"
  dcentry$addDCCreator(creator)
  dcentry$addDCSubject("R")
  dcentry$addDCSubject("FAIR")
  dcentry$addDCSubject("Interoperability")
  dcentry$addDCSubject("Open Science")
  dcentry$addDCDescription("Atom4R offers tools to read/write and publish metadata as Atom XML")
  dcentry$addDCPublisher("GitHub")
  funder <- DCContributor$new(value = "CNRS")
  funder$attrs[["type"]] <- "Funder"
  dcentry$addDCContributor(funder)
  dcentry$addDCRelation("Github repository: https://github.com/eblondel/atom4R")
  dcentry$addDCSource("Atom Syndication format - https://www.ietf.org/rfc/rfc4287")
  dcentry$addDCSource("AtomPub, The Atom publishing protocol - https://tools.ietf.org/html/rfc5023")
  dcentry$addDCSource("Sword API - http://swordapp.org/")
  dcentry$addDCSource("Dublin Core Metadata Initiative - https://www.dublincore.org/")
  dcentry$addDCSource("Guidelines for implementing Dublin Core in XML")
  dcentry$addDCLicense("NONE")
  dcentry$addDCRights("MIT License")
  dcentry$addDCHasPart("part1")
  dcentry$addDCHasPart("part2")
  dcentry$addDCHasVersion("0.2")
  dcentry$addDCIsPartOf("CRAN")
  dcentry$addDCIsPartOf("GitHub")
  dcentry$addDCIsReferencedBy("CRAN")
  dcentry$addDCIsReferencedBy("GitHub")
  dcentry$addDCIsRequiredBy("zen4R")
  dcentry$addDCIsRequiredBy("cloud4R")

  xml <- dcentry$encode()
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}contributor', attribute 'type': The attribute 'type' is not allowed at line 11. 
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}creator', attribute 'affiliation': The attribute 'affiliation' is not allowed at line 14. 
#> [atom4R][WARN] Object 'DCEntry' is INVALID according to Atom XML schemas! 

  #decoding
  dcentry2 <- DCEntry$new(xml = xml)
  xml2 <- dcentry2$encode()
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}contributor', attribute 'type': The attribute 'type' is not allowed at line 11. 
#> [atom4R][WARN] Element '{http://purl.org/dc/terms/}creator', attribute 'affiliation': The attribute 'affiliation' is not allowed at line 14. 
#> [atom4R][WARN] Object 'DCEntry' is INVALID according to Atom XML schemas! 
 # }